// src/application/interfaces/services/media/media-rest/methods/modify-media.interface.ts
import { Media } from 'src/domain/entities/media';

export interface IMediaRestServiceModifyMediaParams {
  indexes: Pick<Media, 'id' | 'user'>;
  data: Partial<
    Pick<
      Media,
      | 'title'
      | 'type'
      | 'description'
      | 'durationSeconds'
      | 'contentBase64'
      | 'available'
    >
  >;
}

export type IMediaRestServiceModifyMediaResult = void;